+++ src/lib/utils/janggi/pieceRules/index.ts
@@ -1,4 +1,3 @@
-// piece rules
 export * from './cannonRules';
 export * from './carRules';
 export * from './elephantRules';
