+++ src/lib/utils/janggi/pieceRules/kingRules.ts
@@ -13,27 +13,27 @@ import {
   DIAG_DY,
 } from '@utils/janggi/constants';
 
-import { getPalaceCenterDiagonalMoves, isInPalace, isTileOccupiedByMyCountry } from '../generalRules';
+import { getPalaceCenterDiagonalMoves, isInPalace, isTileOccupiedByMyCountry } from '../common';
 
-// same as king move
-export const getPossibleScholarMoves = (scholar: Piece, board: Board): Position[] => {
+// same as scholar move
+export const getPossibleKingMoves = (king: Piece, board: Board): Position[] => {
   const possibleMoves: Position[] = [];
-  const { x: currX, y: currY } = scholar.position;
+  const { x: currX, y: currY } = king.position;
 
   // linear move
   for (let i = 0; i < DIRECTION_NUM; i++) {
     const position: Position = new Position(currX + LINEAR_DX[i], currY + LINEAR_DY[i]);
-    if (isInPalace(position) && !isTileOccupiedByMyCountry(scholar.country, position, board)) {
+    if (isInPalace(position) && !isTileOccupiedByMyCountry(king.country, position, board)) {
       possibleMoves.push(position);
     }
   }
 
-  // diagonal move at the corner
+  // diagonal move at the corner of the palace
   for (let i = 0; i < CORNER_NUM; i++) {
-    const isAtPalaceCorner = PALACE_CORNERS[i].some(corner => corner.isSamePosition(scholar.position));
+    const isAtPalaceCorner = PALACE_CORNERS[i].some(corner => corner.isSamePosition(king.position));
     if (isAtPalaceCorner) {
       const diagonalMove: Position = new Position(currX + DIAG_DX[i], currY + DIAG_DY[i]);
-      if (!isTileOccupiedByMyCountry(scholar.country, diagonalMove, board)) {
+      if (!isTileOccupiedByMyCountry(king.country, diagonalMove, board)) {
         possibleMoves.push(diagonalMove);
       }
       break;
@@ -41,7 +41,7 @@ export const getPossibleScholarMoves = (scholar: Piece, board: Board): Position[
   }
 
   // diagonal moves at the center of the palace
-  const palaceCenterDiagonalMoves = getPalaceCenterDiagonalMoves(scholar, board);
+  const palaceCenterDiagonalMoves = getPalaceCenterDiagonalMoves(king, board);
   possibleMoves.push(...palaceCenterDiagonalMoves);
 
   return possibleMoves;
